test: add e2e test for reduceTransparency
